requirements.txt
# LangChain Core
langchain
langchain-core
langchain_community

#Groq
langchain-groq

# Environment Variable Management
python-dotenv

#streamlit
streamlit

pypdf
pymupdf
unstructured
pdfminer.six
unstructured_pytesseract
unstructured_inference

langchain-huggingface
sentence-transformers